requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.2.2
numpy==1.26.4
faiss-cpu==1.8.0.post1
sentence-transformers==3.0.1
scikit-learn==1.5.1
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
orjson==3.10.7
